Removes some redundant classes in tests
For most files there is only one class, with the exception of classes that have a few children
(For example, the formatter classes inherit from a basecase which provides some useful variables,
  and testGateSets2.py depends on a base class in testGateSets.py. This could change if it needed to)
