+++ test/test_packages/objects/testHessian.py
@@ -9,7 +9,8 @@ import sys, os
 
 from ..testutils import BaseTestCase, compare_files, temp_files
 
-class HessianTestCase(BaseTestCase):
+
+class TestHessianMethods(BaseTestCase):
 
     def setUp(self):
         super(HessianTestCase, self).setUp()
@@ -17,7 +18,6 @@ class HessianTestCase(BaseTestCase):
         self.gateset = pygsti.io.load_gateset(compare_files + "/analysis.gateset")
         self.ds = pygsti.objects.DataSet(fileToLoadFrom=compare_files + "/analysis.dataset")
 
-class TestHessianMethods(HessianTestCase):
 
     def test_parameter_counting(self):
         #XY Gateset: SPAM=True
