+++ test/test_packages/optimize/testOptimize.py
@@ -5,6 +5,8 @@ import sys
 import warnings
 import os
 
+from ..testutils import BaseTestCase
+
 def f(x):
     return np.dot(x,x)
 
@@ -14,27 +16,13 @@ def f_vec(x):
 def jac(x):
     return 2*x[None,:]
 
-class OptimizeTestCase(unittest.TestCase):
+class TestOptimizeMethods(BaseTestCase):
 
     def setUp(self):
-        # move working directories
-        self.old = os.getcwd()
-        os.chdir(os.path.abspath(os.path.dirname(__file__)))
-
-        #Set GateSet objects to "strict" mode for testing
-        pygsti.objects.GateSet._strict = True
+        super(TestOptimizeMethods, self).setUp()
 
         self.x0 = np.array( [10,5], 'd')
         self.answer = np.array( [0,0], 'd')
-
-    def tearDown(self):
-        os.chdir(self.old)
-
-    def assertArraysAlmostEqual(self,a,b):
-        self.assertAlmostEqual( np.linalg.norm(a-b), 0 )
-
-
-class TestOptimizeMethods(OptimizeTestCase):
 
     def test_optimize(self):
         old_stdout = sys.stdout
